requirements.txt
reportlab==4.4.7
arabic_reshaper==3.0.0
python-bidi==0.6.6